requirements.txt
# App

Django==1.10.1
djangorestframework==3.4.6
slothauth==0.6.1
django-mobile==0.7.0
celery==3.1.23
django-celery==3.1.17

# Logging

# DevOps

gunicorn==19.6
dj-database-url==0.4.1
django-storages==1.5.0
django-boto==0.3.9

# Testing

factory_boy==2.7.0

# Utility

django-extensions==1.7.2
Werkzeug==0.11.10
PyYAML==3.11
